+++ ErsatzTV.Core/Scheduling/RandomizedMediaCollectionEnumerator.cs
@@ -15,7 +15,7 @@ public class RandomizedMediaCollectionEnumerator : IMediaCollectionEnumerator
     {
         _mediaItems = mediaItems;
         _lazyMinimumDuration =
-            new Lazy<Option<TimeSpan>>(() => _mediaItems.Bind(i => i.GetDuration()).OrderBy(identity).HeadOrNone());
+            new Lazy<Option<TimeSpan>>(() => _mediaItems.Bind(i => i.GetNonZeroDuration()).OrderBy(identity).HeadOrNone());
         _random = new Random(state.Seed);
 
         State = new CollectionEnumeratorState { Seed = state.Seed };