+++ ErsatzTV.Core/Scheduling/ShuffleInOrderCollectionEnumerator.cs
@@ -34,7 +34,7 @@ public class ShuffleInOrderCollectionEnumerator : IMediaCollectionEnumerator
         _random = new Random(state.Seed);
         _shuffled = Shuffle(_collections, _random);
         _lazyMinimumDuration =
-            new Lazy<Option<TimeSpan>>(() => _shuffled.Bind(i => i.GetDuration()).OrderBy(identity).HeadOrNone());
+            new Lazy<Option<TimeSpan>>(() => _shuffled.Bind(i => i.GetNonZeroDuration()).OrderBy(identity).HeadOrNone());
 
         State = new CollectionEnumeratorState { Seed = state.Seed };
         while (State.Index < state.Index)